I had a lot of trouble finishing up the param server. I wanted to have a callback be registered for all param changes, independent of the param name.
Using the async mechanism provided by the .proto files we can register a callback. When the callback is invoked in the param server implementation the program segfaults. For now I commented out the code that calls the callback. I figured it was a problem with an incorrect thread accessing the data, but the program would then crash if I added a std::mutex as a member variable. This left me stumped, so I'm leaving it there. If the callback can be called,then the implementation for detecting params changed will be done. This can then be published out to a topic via a ros publisher as a pair of string and value.